refactor: simplify CLI command handlers with command pattern
Break down the large search function into focused, reusable components:
- Create CommandHandler trait for standardized command processing
- Extract SearchCommandHandler with specialized methods for query building,
- Remove search-related functions and move logic to dedicated handler
